--- scrapyselectors/__init__.py
@@ -1,4 +1,0 @@
-"""
-Selectors
-"""
-from scrapyselectors.unified import *
